rappi-insights/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
anthropic==0.39.0
openpyxl==3.1.5
pandas==2.2.3
numpy==2.1.1
python-dotenv==1.0.1
pydantic==2.9.2
markdown==3.7
jinja2==3.1.4
